xss_scanner/requirements.txt
# requirements.txt
requests
beautifulsoup4
aiohttp
playwright
pytest
argpars
colorama
PyJWT